Get rid of `inventory`
