--- packages/vm-derive-impl/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "cosmwasm-vm-derive-impl"
-edition = "2021"
-version.workspace = true
-
-[lib]
-proc-macro = true
-
-[dependencies]
-blake3 = "1.5.4"
-proc-macro2 = "1.0.86"
-quote = "1.0.37"
-syn = { version = "2.0.77", features = ["extra-traits", "full"] }
