+++ packages/vm-derive/Cargo.toml
@@ -3,6 +3,11 @@ name = "cosmwasm-vm-derive"
 edition = "2021"
 version.workspace = true
 
+[lib]
+proc-macro = true
+
 [dependencies]
-cosmwasm-vm-derive-impl = { version = "=2.1.3", path = "../vm-derive-impl" }
-inventory = "0.3.15"
+blake3 = "1.5.4"
+proc-macro2 = "1.0.86"
+quote = "1.0.37"
+syn = { version = "2.0.77", features = ["extra-traits", "full"] }
