+++ packages/vm/src/wasm_backend/mod.rs
@@ -7,4 +7,4 @@ mod limiting_tunables;
 pub use engine::make_compiler_config;
 
 pub use compile::compile;
-pub use engine::{make_compiling_engine, make_runtime_engine};
+pub use engine::{make_compiling_engine, make_runtime_engine, COST_FUNCTION_HASH};
